Split shared hooks and move them to navigation/hooks
