+++ src/app/navigation/hooks/closeModalHook.ts
@@ -1,0 +1,3 @@
+import { AppServices, MiddlewareAPI } from '../../types';
+
+export const closeModal = (services: MiddlewareAPI & AppServices) => services.history.goBack;
