+++ src/app/navigation/hooks/openModalHook.spec.ts
@@ -1,17 +1,17 @@
-import createTestServices from '../../../../test/createTestServices';
-import createTestStore from '../../../../test/createTestStore';
-import { book as archiveBook, page } from '../../../../test/mocks/archiveLoader';
-import { mockCmsBook } from '../../../../test/mocks/osWebLoader';
-import { MiddlewareAPI, Store } from '../../../types';
-import { initialState as initialContentState } from '../../reducer';
-import { formatBookData } from '../../utils';
+import createTestServices from '../../../test/createTestServices';
+import createTestStore from '../../../test/createTestStore';
+import { book as archiveBook, page } from '../../../test/mocks/archiveLoader';
+import { mockCmsBook } from '../../../test/mocks/osWebLoader';
+import { initialState as initialContentState } from '../../content/reducer';
+import { formatBookData } from '../../content/utils';
+import { MiddlewareAPI, Store } from '../../types';
 
 const book = formatBookData(archiveBook, mockCmsBook);
 
 describe('openModal', () => {
   let store: Store;
   let helpers: ReturnType<typeof createTestServices> & MiddlewareAPI;
-  let hookFactory: typeof import ('./sharedHooks').openModal;
+  let hookFactory: typeof import ('./openModalHook').openModal;
   let dispatch: jest.SpyInstance;
 
   beforeEach(() => {
@@ -35,7 +35,7 @@ describe('openModal', () => {
 
     dispatch = jest.spyOn(helpers, 'dispatch');
 
-    hookFactory = (require('./sharedHooks').openModal);
+    hookFactory = (require('./openModalHook').openModal);
   });
 
   it('sets correct location', () => {
